fix: :white_check_mark: Re-added volunteer tests
